Use type from compile time value for binary expression

+++ compiler/testData/loadKotlin/annotations/classes/WithMultipleArguments.kt
@@ -2,4 +2,4 @@ package test
 
 annotation class Anno(val int: Int, val string: String, val double: Double)
 
-Anno(42, "OK", 3.14) class Class
+Anno(42.toInt(), "OK", 3.14.toDouble()) class Class
